assignment_4/requirements.txt
pyghidra
networkx
python-dotenv
fastapi
uvicorn[standard]
matplotlib